# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "North Bound Interface (NBI) REST API exposed by voltha-go-controller (VGC).\nIt provides configuration and operational endpoints for subscribers, bandwidth\nprofiles, IGMP proxy, multicast, flows, devices, ports, meters, groups, DHCP\nallocations, MAC learning and the OLT flow service.",
        "title": "VOLTHA Go Controller NBI REST API",
        "contact": {
            "name": "Open Networking Foundation",
            "url": "https://opennetworking.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/vgc/v1",
    "paths": {
...
        "/igmp-proxy": {
            "post": {
                "description": "Configure IGMP proxy for a multicast VLAN.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IGMP Proxy"
                ],
                "summary": "Add IGMP proxy configuration",
                "parameters": [
                    {
                        "description": "IGMP proxy configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nbi.IgmpProxy"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IGMP proxy configuration accepted"
                    },
                    "500": {
                        "description": "Failed to read request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/igmp-proxy/{outgoingigmpvlanid}": {
            "delete": {
                "description": "Remove IGMP proxy configuration for the given outgoing IGMP VLAN id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IGMP Proxy"
                ],
                "summary": "Delete IGMP proxy configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Outgoing IGMP VLAN id",
                        "name": "outgoingigmpvlanid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IGMP proxy configuration deleted"
                    }
                }
            }
        },
...
        "/olt/{serialNumber}": {
            "get": {
                "description": "Retrieve the device configuration for a specific OLT by serial number.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Configuration"
                ],
                "summary": "Get device configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OLT serial number",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.DeviceConfigPayload"
                        }
                    },
                    "400": {
                        "description": "Device not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add or update the device configuration for an OLT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Configuration"
                ],
                "summary": "Add device configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OLT serial number",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.DeviceConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device configuration added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/updateDevice/{deviceId}": {
            "post": {
                "description": "Update the uplink configuration for a device.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Configuration"
                ],
                "summary": "Update device uplink configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device configuration with updated NNI ports",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.DeviceConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device configuration updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "application.DeviceConfig": {
            "type": "object",
            "properties": {
                "hardwareIdentifier": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "nasId": {
                    "type": "string"
                },
                "nniDhcpTrapVid": {
                    "type": "integer"
                },
                "nniPorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uplinkPort": {
                    "type": "string"
                }
            }
        },
...
        "nbi.IgmpProxy": {
            "type": "object",
            "properties": {
                "enableigmpprovisioning": {
                    "type": "string"
                },
                "fastleave": {
                    "type": "string"
                },
                "globalconnectpoint": {
                    "type": "string"
                },
                "globalconnectpointmode": {
                    "type": "string"
                },
                "igmpcos": {
                    "type": "integer"
                },
                "igmpunicos": {
                    "type": "integer"
                },
                "keepalivecount": {
                    "type": "integer"
                },
                "keepaliveinterval": {
                    "type": "integer"
                },
                "lastqueryinterval": {
                    "type": "integer"
                },
                "maxresp": {
                    "type": "integer"
                },
                "outgoingigmpinnervlanid": {
                    "type": "integer"
                },
                "outgoingigmpvlanid": {
                    "type": "integer"
                },
                "outgoingigmpwithv3": {
                    "type": "string"
                },
                "periodicquery": {
                    "type": "string"
                },
                "requestdsigmppackets": {
                    "type": "boolean"
                },
                "sourcedeviceandport": {
                    "type": "string"
                }
            }
        },
...
        "onosnbi.DeviceConfigPayload": {
            "type": "object",
            "properties": {
                "deviceConfig": {
                    "$ref": "#/definitions/application.DeviceConfig"
                }
            }
        },
...
    }
}